backend/services/database_service/requirements.txt
pydantic
prisma
python-dotenv
fastapi
uvicorn